+ Rename directory http_handler
