+++ lib/http-handler/delete.js
